Very raw, but working version of rust minimax compatible with Python
