--- experiments/test_maturin.py
@@ -1,3 +1,0 @@
-from rs_minimax import sum_as_string
-
-print(sum_as_string(1, 2))
